# src/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Violeta App BE",
    "description": "Documentação da API do backend do Aplicativo Violeta",
    "version": "0.1.1"
  },
  "servers": [
    {
      "url": "http://localhost:8989/",
      "description": "Server local"
    }
  ],
  "paths": {
    "/occurrences": {
      "get": {
        "tags": ["Occurrences"],
        "summary": "Retorna todas as ocorrências.",
        "responses": {
          "200": {
            "description": "Array JSON com todas as ocorrências.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Occurrence"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Occurrences"],
        "summary": "Cria novas ocorrências.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/OccurrenceWithoutId"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ocorrências criadas com sucesso."
          },
          "400": {
            "description": "Requisição inválida."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/occurrences/{occurrence_id}": {
      "get": {
        "tags": ["Occurrences"],
        "summary": "Retorna uma ocorrência pelo ID.",
        "parameters": [
          {
            "name": "occurrence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados da ocorrência.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Occurrence"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Occurrences"],
        "summary": "Atualiza uma ocorrência pelo ID.",
        "parameters": [
          {
            "name": "occurrence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Occurrence"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ocorrência atualizada."
          }
        }
      },
      "delete": {
        "tags": ["Occurrences"],
        "summary": "Deleta uma ocorrência pelo ID.",
        "parameters": [
          {
            "name": "occurrence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ocorrência deletada com sucesso."
          }
        }
      }
    },
    "/occurrences/{user_id}": {
      "get": {
        "tags": ["Occurrences"],
        "summary": "Retorna a(s) ocorrência(s) dado um user_id.",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados da(s) ocorrência(s).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Occurrence"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Retorna todas as usuárias.",
        "responses": {
          "200": {
            "description": "Array JSON com todas as usuárias.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Cria novas usuárias.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/UserWithoutId"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuárias criadas com sucesso."
          },
          "400": {
            "description": "Requisição inválida."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Retorna uma usuária pelo ID.",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados da usuária.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Atualiza uma usuária pelo ID.",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuária atualizada."
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Deleta uma usuária pelo ID.",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuária deletada com sucesso."
          }
        }
      }
    },
    "/locais_seguros": {
      "get": {
        "tags": ["Locais Seguros"],
        "summary": "Retorna todos os locais seguros.",
        "responses": {
          "200": {
            "description": "Array JSON com todos os locais seguros.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocalSeguro"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Locais Seguros"],
        "summary": "Cria novos locais seguros.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/LocalSeguroWithoutId"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Locais seguros criados com sucesso."
          },
          "400": {
            "description": "Requisição inválida."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/locais_seguros/{id}": {
      "get": {
        "tags": ["Locais Seguros"],
        "summary": "Retorna um local seguro pelo ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do local seguro.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalSeguro"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Locais Seguros"],
        "summary": "Atualiza um local seguro pelo ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalSeguro"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Local Seguro atualizado."
          }
        }
      },
      "delete": {
        "tags": ["Locais Seguros"],
        "summary": "Deleta um local seguro pelo ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Local seguro deletado com sucesso."
          }
        }
      }
    },
    "/alertas": {
      "get": {
        "tags": ["Alertas"],
        "summary": "Retorna todas os alertas.",
        "responses": {
          "200": {
            "description": "Array JSON com todas os alertas.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Alerta"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Alertas"],
        "summary": "Cria novos alertas.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AlertaWithoutId"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Alertas criados com sucesso."
          },
          "400": {
            "description": "Requisição inválida."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/alertas/{alerta_id}": {
      "get": {
        "tags": ["Alertas"],
        "summary": "Retorna um alerta pelo ID.",
        "parameters": [
          {
            "name": "alerta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do alerta.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Alerta"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Alertas"],
        "summary": "Atualiza um alerta pelo ID.",
        "parameters": [
          {
            "name": "alerta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Alerta"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Alerta atualizado."
          }
        }
      },
      "delete": {
        "tags": ["Alertas"],
        "summary": "Deleta um alerta pelo ID.",
        "parameters": [
          {
            "name": "alerta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alerta deletado com sucesso."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Occurrence": {
        "type": "object",
        "properties": {
          "occurrence_id": {
            "type": "string"
          },
          "occurrence_description": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "date": {
            "type": "string"
          },
          "time": {
            "type": "string"
          },
          "main_reason": {
            "type": "string"
          },
          "victim_situation": {
            "type": "string"
          },
          "anonymous": {
            "type": "boolean"
          },
          "victim_name": {
            "type": "string"
          },
          "occurrence_score": {
            "type": "integer"
          }
        }
      },
      "OccurrenceWithoutId": {
        "type": "object",
        "properties": {
          "occurrence_description": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "date": {
            "type": "string"
          },
          "time": {
            "type": "string"
          },
          "main_reason": {
            "type": "string"
          },
          "victim_situation": {
            "type": "string"
          },
          "anonymous": {
            "type": "boolean"
          },
          "victim_name": {
            "type": "string"
          },
          "occurrence_score": {
            "type": "integer"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string"
          },
          "user_name": {
            "type": "string"
          },
          "user_login": {
            "type": "string"
          },
          "user_password": {
            "type": "string"
          }
        }
      },
      "UserWithoutId": {
        "type": "object",
        "properties": {
          "user_name": {
            "type": "string"
          },
          "user_login": {
            "type": "string"
          },
          "user_password": {
            "type": "string"
          }
        }
      },
      "Alerta": {
        "type": "object",
        "properties": {
          "alerta_id": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "upvotes": {
            "type": "integer"
          },
          "downvotes": {
            "type": "integer"
          }
        }
      },
      "AlertaWithoutId": {
        "type": "object",
        "properties": {
          "descricao": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          }
        }
      },
      "LocalSeguro": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "tipo": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          }
        }
      },
      "LocalSeguroWithoutId": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "tipo": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          }
        }
      }
    }
  }
}